Move shared test code to class to allow future SIAv2 testing
